Fix enable_snat column migration for Postgres + SQLite
Previously, the column was being added with NOT NULL by default.
This caused an issue when performing a migration in Postgres and
SQLite where the default value was not being set. Instead, this change
allows NULL values on creation, sets all values to the default,
and then sets NOT NULL.

Closes-Bug: 1241577

Change-Id: I10039130042796177d26afdec11d08a5ef0be18c